feat: renamed auth0 service to userInfo service
affects: tymly, tymly-auth-auth0-plugin

auth0 was badly named because it was providing things which weren't specific to Auth0, and because
it wasn't actually doing any authentication. Right now, the only method used is providing an email
address from a userId.  So I chose userInfo as the name, because that's what it seems to be doing :)