fix: Moved resolveInputPaths into find-base from statebox
This function was part of the content the Task state passes in to a state resource. However, it was
only used in this one place, and there's no prospect of it being used anywhere else, so the code
should live here. Technically this is a refactoring, but because it involves moving code across
packages I've marked it as a bug.